ClusterBench variants that passivate to an external store
This commit adds both EE6 and EE7 variant of the passivating
webapp that offloads to an external cache store. To be able to use
the external store safely, a cache that is used exclusively
by the application is required. Such cache must be configured
in jboss-web.xml.

The changes are twofold:

1. An extra WAR is added (see clusterbench-eeN-web/webapp-offload)
   that configures <cache-name> in jboss-web.xml. These are copied
   directly from clusterbench-eeN-web/webapp-passivating and
   modified accordingly.
2. For EE6, an extra EAR is created that contains the offload WAR.
   This is because for EE6, the pasivating variant is also built
   as an extra EAR. For EE7, the EAR now includes the offload WAR,
   again because the passivating variant is also added to main EAR.

The "test" in the EE7 EAR module is also removed, because it relies
on default application server configuration, but the offload WAR
requires a modification.